requirements.txt
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
httpx>=0.26.0,<1.0
pydantic>=2.0,<3.0
python-dotenv>=1.0.0,<2.0
faiss-cpu>=1.7.0,<2.0
sentence-transformers>=2.0.0,<3.0
numpy>=1.24.0,<3.0
streamlit>=1.28.0,<2.0
plotly>=5.18.0,<6.0
pandas>=2.0.0,<3.0
requests>=2.28.0,<3.0
